Crate webview_app

webview_app

Integration of a web view in an application window like Electron, but using rust as programming language. On Windows WebView2 will be used as web view, on Linux it is WebKitGTK.

A simple example to create and run a simple web view app displaying crates homepage:

use webview_app::{application::Application, webview::WebView};
 
fn on_activate(app: &Application)->WebView {
    WebView::builder(app)
        .title("Rust Web View 🦞")
        .url("https://crates.io/crates")
        .default_contextmenu_disabled()
        .build();
}
 
fn main() {
    Application::new("de.uriegel.hello")
    .on_activate(on_activate)
    .run();
}

A tutorial for using webview_app:

https://uriegel.github.io/webview-app/doc/tutorial/

Modules

application

request

webview

This module contains all the important structs and implementations to create, configure and run a webview window.